requirements.txt
discord_py_interactions==5.12.1
Requests==2.31.0
discord.py-interactions[voice]